Исправления и улучшения
