+++ src/main/java/ru/yandex/practicum/filmorate/dao/extractor/GenreExtractor.java
@@ -1,9 +1,9 @@
-package ru.yandex.practicum.filmorate.dal.extractor;
+package ru.yandex.practicum.filmorate.dao.extractor;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.ResultSetExtractor;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.model.Mpa;
+import ru.yandex.practicum.filmorate.model.Genre;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -11,22 +11,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Component
-public class MpaExtractor implements ResultSetExtractor<List<Mpa>> {
+public class GenreExtractor implements ResultSetExtractor<List<Genre>> {
 
 	@Override
-	public List<Mpa> extractData(ResultSet rs)
+	public List<Genre> extractData(ResultSet rs)
 			throws SQLException, DataAccessException {
-		List<Mpa> mpas = new ArrayList<>();
+		List<Genre> genres = new ArrayList<>();
 
 		//итерируемся по всем строкам ResultSet
 		while (rs.next()) {
-			Mpa mpa = new Mpa();
-			mpa.setId(rs.getInt("MPA_ID"));
-			mpa.setName(rs.getString("NAME"));
+			Genre genre = new Genre();
+			genre.setId(rs.getInt("GENRE_ID"));
+			genre.setName(rs.getString("NAME"));
 
-			mpas.add(mpa);
+			genres.add(genre);
 		}
 
-		return mpas;
+		return genres;
 	}
 }
