+++ src/test/java/ru/yandex/practicum/filmorate/FilmorateApplicationTests.java
@@ -7,14 +7,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.test.context.ContextConfiguration;
-import ru.yandex.practicum.filmorate.dal.DirectorRepository;
-import ru.yandex.practicum.filmorate.dal.EventRepository;
-import ru.yandex.practicum.filmorate.dal.FilmRepository;
-import ru.yandex.practicum.filmorate.dal.UserRepository;
-import ru.yandex.practicum.filmorate.dal.extractor.DirectorExtractor;
-import ru.yandex.practicum.filmorate.dal.extractor.EventExtractor;
-import ru.yandex.practicum.filmorate.dal.extractor.FilmExtractor;
-import ru.yandex.practicum.filmorate.dal.extractor.UserExtractor;
+import ru.yandex.practicum.filmorate.dao.DirectorRepository;
+import ru.yandex.practicum.filmorate.dao.EventRepository;
+import ru.yandex.practicum.filmorate.dao.FilmRepository;
+import ru.yandex.practicum.filmorate.dao.UserRepository;
+import ru.yandex.practicum.filmorate.dao.extractor.DirectorExtractor;
+import ru.yandex.practicum.filmorate.dao.extractor.EventExtractor;
+import ru.yandex.practicum.filmorate.dao.extractor.FilmExtractor;
+import ru.yandex.practicum.filmorate.dao.extractor.UserExtractor;
 import ru.yandex.practicum.filmorate.dto.UserDto;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
